extracts Fieldset component
